+++ src/main/java/webserver/http/request/handler/get/UserCreateByQueryParamRequestHandler.java
@@ -1,8 +1,8 @@
-package webserver.http.request.handler.post;
+package webserver.http.request.handler.get;
 
 import webserver.http.request.handler.RequestHandler;
 
-public class UserCreatePostRequestHandler implements RequestHandler {
+public class UserCreateByQueryParamRequestHandler implements RequestHandler {
     private static final String REQUEST_INDEX = "/user/create";
 
     public static String requestIndex() {
