+++ src/main/java/webserver/http/response/handler/PostResponseGroup.java
@@ -2,7 +2,7 @@ package webserver.http.response.handler;
 
 import java.util.Map;
 
-import webserver.http.request.handler.post.UserCreatePostRequestHandler;
+import webserver.http.request.handler.post.UserCreateByBodyRequestHandler;
 import webserver.http.request.handler.post.UserLoginRequestHandler;
 import webserver.http.request.handler.exception.InvalidRequestException;
 import webserver.http.response.handler.post.UserCreatePostResponseHandler;
@@ -10,7 +10,7 @@ import webserver.http.response.handler.post.UserLoginResponseHandler;
 
 public class PostResponseGroup implements ResponseGroup {
     private final Map<String, ResponseHandler> handlerMap = Map.ofEntries(
-            Map.entry(UserCreatePostRequestHandler.requestIndex(), new UserCreatePostResponseHandler()),
+            Map.entry(UserCreateByBodyRequestHandler.requestIndex(), new UserCreatePostResponseHandler()),
             Map.entry(UserLoginRequestHandler.requestIndex(), new UserLoginResponseHandler())
     );
